Add support for Visual Studio 2017
Upgrage the .vsix manifest to version 3.0, which is required for Visual Studio 2017 support. This drops support for Visual Studio 2010, but this is the recommended approach by Microsoft. If such support is needed, new extension with new Id should be provided.

Add some references as packages, so that the project could be build on machines where older versions of Visual Studio are not installed.

The source code is now only buildable using Visual Studio 2017, which is also limitation of the new VS extensions approach.